Rewrite filling in hashed password using browser.runtime.executeScript()
Because we are using page actions to show the hash dialog we can simply do
browser.runtime.executeScript() on the active tab of the current window to
fill in the hash word in the last focused input.

This reduces the background script by a lot, and also means we don't have to
keep track of focusing state in the content script.

Because we fill in the password field with executeScript(), we do not need a
a connection between the content script and the background script, removing
the need for ports or browser.runtime.sendMessage()/onMessage.